fix: Parse single-parameter function literals correctly.
After advancing twice (once if expectPeek(.LeftParen) function was
successful, and then again to move the parser past the expected left
paren), the if expression only checked the state of the peek token. For
function literals in the form `fn(x){...}`, at this point, the parser
would see a right paren as the peek token, thus failing to parse the
parameter. This caused the exepectPeek(.LeftBrace) statement to error,
causing the program to panic.