parser: Fix zero argument function literal parsing.
Previously the parser would fail when parsing function literals such as
(`fn(){...}`) because the parser would already be advanced to the right
parenthesis token by the time we called `expectPeek(.r_paren)`. Now we
check for the right paren token when we encounter our while loop stop
condition when parsing the function parameters.