evaluator: Evaluate array literals and index expressions. parser: Fix parseExpression skipping over semicolon token for consecutive expressions.
The while loop condition in the parseExpression method now checks if the
__current token__ is a semicolon instead of the peek token, preventing
the parser from accidentally continuing to parse expressions and
statements separated by a semicolon.

e.g.,
```
    >> let double = fn(x) { x * 2 }; [1, double(2), 3]
```
Previously, in this example, the parser first parses the let statement.
After parsing the expression (`fn(x) { x * 2 }`) the current token at
that point is a semicolon. Since the parser checked the peek token
instead of current token, it interprets the expression as the left hand
side of some __infix expression__, and attempts to parse the rest of
tokens accordingly by calling parseInfixToken. parseInfixToken advnances
the parser, at which point it encounters the left bracket token thinking
the most recently parsed expression is part of a index exprssion,
causing the parser to fail.